agrego NoTurno como opción
