+++ materias/templates/materias/cambiar_una_carga_publicada.html
@@ -5,11 +5,11 @@
 <link rel="stylesheet" type="text/css" href="{% static 'materias/horarios.css' %}">
 <link rel="stylesheet" type="text/css" href="{% static 'materias/administrar.css' %}">
 
-<h1>Cambiar la carga docente de {{ carga.docente }} a {{ carga.turno }}</h1>
+<h2>Cambiar la carga docente de {{ carga.docente }} a {{ carga.turno }}</h2>
 <form action="{% url 'materias:cambiar_una_carga_publicada' carga.id %}" method="post">
     {% csrf_token %}
 
-    <h2>Turno</h2>
+    <h3>Turno</h3>
     <ul>
         <li>
             Actual: {{ carga.turno }}
